refactor powermodels files and added to converter :sheep:
