+++ pandapower/converter/__init__.py
@@ -1,3 +1,4 @@
 from pandapower.converter.matpower import *
 from pandapower.converter.pypower import *
+from pandapower.converter.powermodels import *
