+++ pandapower/converter/powermodels/__init__.py
@@ -1,0 +1,1 @@
+from pandapower.converter.powermodels.to_pm import *
