+++ pandapower/runpm.py
@@ -8,10 +8,9 @@ import os
 
 from pandapower import pp_dir
 from pandapower.auxiliary import _add_ppc_options, _add_opf_options
-from pandapower.opf.pm_conversion import _runpm
-from pandapower.opf.pm_ots import read_ots_results
+from pandapower.opf.run_powermodels import _runpm
+from pandapower.converter.powermodels.from_pm import read_ots_results, read_tnep_results
 from pandapower.opf.pm_storage import add_storage_opf_settings, read_pm_storage_results
-from pandapower.opf.pm_tnep import read_tnep_results
 
 
 def runpm(net, julia_file, pp_to_pm_callback=None, calculate_voltage_angles=True,
